BT_HAL: Fix tautological compare warning in uart_controller
Issue Description:
Compiler warning regarding tautological comparison.

Root Cause:
IBS_WAKE_IND (0xFD) is outside the valid range of 'enum HciPacketType'.
Comparing an enum variable directly with this value is considered
always false by the compiler.

Solution:
Explicitly cast 'hci_packet_type_' to uint8_t to enforce a raw byte
comparison.

+++ hidl_transport/bt/1.0/default/uart_controller.cpp
@@ -1586,7 +1586,7 @@ void UartController::OnDataReady(int fd)
         if (invalid_bytes_counter_ < TX_RX_PKT_ASC_SIZE)
           inv_bytes[invalid_bytes_counter_] = hci_packet_type_;
         invalid_bytes_counter_++;
-        if (hci_packet_type_ == IBS_WAKE_IND) {
+        if (static_cast<uint8_t>(hci_packet_type_) == IBS_WAKE_IND) {
           sync_bytes_rcvd_++;
           ALOGE("%s: OutOfSync - Sync Byte Rcvd:%d", __func__, sync_bytes_rcvd_);
           if (sync_bytes_rcvd_ == MIN_BYTES_FOR_SYNC_CONFIRM) {
